backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.28.0
python-multipart>=0.0.18
PyJWT>=2.8.0
pgvector>=0.3.0
sentence-transformers>=3.0.0
minio>=7.2.0
bcrypt>=4.2.0
alembic>=1.13.0
email-validator>=2.2.0
sentry-sdk[fastapi]>=2.16.0
